refactor(arrow): switching the slice utilities to use int instead of int64
Using `int` is more common than `int64` for lengths so, even if arrow
uses `int64`, the slice lengths should still use `int`. Additionally,
the arrow `Len()` method for slices uses `int` as the return value so
this is better for that anyway.

+++ arrow/string.go
@@ -18,8 +18,8 @@ func NewString(vs []string, alloc *memory.Allocator) *array.Binary {
 	return a
 }
 
-func StringSlice(arr *array.Binary, i, j int64) *array.Binary {
-	data := array.NewSliceData(arr.Data(), i, j)
+func StringSlice(arr *array.Binary, i, j int) *array.Binary {
+	data := array.NewSliceData(arr.Data(), int64(i), int64(j))
 	defer data.Release()
 	return array.NewBinaryData(data)
 }
